Refactor common functions that don't depend on DOM APIs to seperate utils lib. Allow invoking from build-locale-data that operates in node
