+++ lib/dateTime.js
@@ -1,6 +1,7 @@
-import { getDocumentLocaleSettings, merge } from './common.js';
 import { formatDateTimeSkeleton } from './dateTimeSkeleton.js';
+import { getDocumentLocaleSettings } from './common.js';
 import { localeData } from './locale-data/current.js';
+import { merge } from './common.util.js';
 import { resolveSupportedLocale } from './locale-data/supported.js';
 
 // timezone abbreviations and offsets from https://www.timeanddate.com/time/zones/
