+++ lib/fileSize.js
@@ -1,5 +1,6 @@
-import { getLanguage, validateFormatValue } from './common.js';
 import { formatNumber } from './number.js';
+import { getLanguage } from './common.js';
+import { validateFormatValue } from './common.util.js';
 
 export function getFileSizeDescriptor() {
