+++ lib/number.js
@@ -1,9 +1,8 @@
 import {
 	getDocumentLocaleSettings,
-	getLanguage,
-	merge,
-	validateFormatValue
+	getLanguage
 } from './common.js';
+import { merge, validateFormatValue } from './common.util.js';
 
 function formatPositiveInteger(value, descriptor, useGrouping) {
 	value = Math.floor(value);
